Extract the `DPsub` join-order optimizer
